Enum rubble::att::pdus::Opcode

pub enum Opcode {

    ErrorRsp,
    ExchangeMtuReq,
    ExchangeMtuRsp,
    FindInformationReq,
    FindInformationRsp,
    FindByTypeValueReq,
    FindByTypeValueRsp,
    ReadByTypeReq,
    ReadByTypeRsp,
    ReadReq,
    ReadRsp,
    ReadBlobReq,
    ReadBlobRsp,
    ReadMultipleReq,
    ReadMultipleRsp,
    ReadByGroupReq,
    ReadByGroupRsp,
    WriteReq,
    WriteRsp,
    WriteCommand,
    SignedWriteCommand,
    PrepareWriteReq,
    PrepareWriteRsp,
    ExecuteWriteReq,
    ExecuteWriteRsp,
    HandleValueNotification,
    HandleValueIndication,
    HandleValueConfirmation,
    Unknown(u8),

}

Specifies an ATT operation to perform.

The byte values assigned to opcodes are chosen so that the most significant 2 bits indicate additional information that can be useful in some cases:

MSb                            LSb
+-----------+---------+----------+
| Signature | Command |  Method  |
|   1 bit   |  1 bit  |  6 bits  |
+-----------+---------+----------+

Signature is set to 1 to indicate that the Attribute Opcode and Parameters are followed by an Authentication Signature. This is only allowed for the Write Command, resulting in the SignedWriteCommand.

Command is set to 1 when the PDU is a command. This is done purely so that the server can ignore unknown commands. Unlike Requests, Commands are not followed by a server response.

Method defines which operation to perform.

Variants

ErrorRsp

ExchangeMtuReq

ExchangeMtuRsp

FindInformationReq

FindInformationRsp

FindByTypeValueReq

FindByTypeValueRsp

ReadByTypeReq

ReadByTypeRsp

ReadReq

ReadRsp

ReadBlobReq

ReadBlobRsp

ReadMultipleReq

ReadMultipleRsp

ReadByGroupReq

ReadByGroupRsp

WriteReq

WriteRsp

WriteCommand

SignedWriteCommand

PrepareWriteReq

PrepareWriteRsp

ExecuteWriteReq

ExecuteWriteRsp

HandleValueNotification

HandleValueIndication

HandleValueConfirmation

Unknown(u8)

Implementations

impl Opcode

pub fn raw(&self) -> u8

Returns the raw byte corresponding to the opcode self.

pub fn is_authenticated(&self) -> bool

Returns whether the Signature bit in this opcode is set.

If the bit is set, this is an authenticated operation. The opcode parameters are followed by a 12-Byte signature.

pub fn is_command(&self) -> bool

Returns whether the Command bit in this opcode is set.

Commands sent to the server are not followed by a server response (ie. it is not indicated whether they succeed). Unimplemented commands should be ignored, according to the spec.