Struct rubble::att::AttributeServer

pub struct AttributeServer<A: AttributeProvider> {
    attrs: A,
}

An Attribute Protocol server providing read and write access to stored attributes.

Fields
attrs: A
Implementations

impl<A: AttributeProvider> AttributeServer<A>

pub fn new(attrs: A) -> Self

Creates an AttributeServer hosting attributes from an AttributeProvider.

pub fn with_sender<'a>(
&'a mut self,
sender: Sender<'a>
) -> AttributeServerTx<'a, A>

Prepares for performing a server-initiated action (eg. sending a notification/indication).

The caller must ensure that sender has at least RSP_PDU_SIZE bytes of free space available.

It is usually not necessary to use this function. Instead, call L2CAPStateTx::att.

pub fn provider(&mut self) -> &mut A

Provides mutable access to the underlying AttributeProvider.

fn att_mtu(&self) -> u8

Returns the ATT_MTU value, the maximum size of an ATT PDU that can be processed and sent out by the server.

fn process_request(
&mut self,
msg: &AttPdu<'_>,
responder: &mut Sender<'_>
) -> Result<(), AttError>

Process an incoming request (or command) PDU and return a response.

This may return an AttError, which the caller will then send as a response. In the success case, this method will send the response (if any).