Trait rubble::bytes::FromBytes

pub trait FromBytes<'a>: Sized {
    // Required method
    fn from_bytes(bytes: &mut ByteReader<'a>) -> Result<Self, Error>;
}

Trait for decoding values from a byte slice.

Required Methods

fn from_bytes(bytes: &mut ByteReader<'a>) -> Result<Self, Error>

Decode a Self from a byte slice, advancing bytes to point past the data that was read.

If bytes contains data not valid for the target type, or contains an insufficient number of bytes, an error will be returned and the state of bytes is unspecified (it can point to arbitrary data).