Trait rubble::bytes::RawRepr

pub trait RawRepr<T> {
    // Required methods
    fn from_raw(raw: T) -> Self;
    fn as_raw(&self) -> T;
}

Indicates that a type can be cheaply converted from and to a raw representation of type T.

This is meant to be used with Field, and allows representing types as their raw contents in memory, while allowing ergonomic conversion to their typed representation.

Required Methods

fn from_raw(raw: T) -> Self

fn as_raw(&self) -> T