Trait rubble::config::Config

pub trait Config {
    type Timer: Timer;
    type Transmitter: Transmitter;
    type ChannelMapper: ChannelMapper;
    type PacketQueue: PacketQueue;
}

Trait for Rubble stack configurations.

This trait defines a number of types to be used throughout the layers of the BLE stack, which define capabilities, data structures, data, and hardware interface types to be used.

Every application must define a type implementing this trait and supply it to the stack.

Required Associated Types

type Timer: Timer

A time source with microsecond resolution.

type Transmitter: Transmitter

The BLE packet transmitter (radio).

type ChannelMapper: ChannelMapper

The L2CAP channel mapper in use.

This type also provides access to the attributes hosted by the ATT server.

type PacketQueue: PacketQueue

The packet queue to use for exchanging data between the real-time Link-Layer and non-realtime parts of the stack.