Trait rubble::l2cap::ChannelMapper

pub trait ChannelMapper {
    type AttributeProvider: AttributeProvider;
    fn lookup(
        &mut self, 
        channel: Channel
    ) -> Option<ChannelData<'_, dyn ProtocolObj>>;

    fn att(
        &mut self
    ) -> ChannelData<'_, AttributeServer<Self::AttributeProvider>>;
}

Trait for L2CAP channel mappers that provide access to the protocol or service behind a CID.

Associated Types

type AttributeProvider: AttributeProvider

The attribute provider used by the ATT server.

Loading content...
Required methods

fn lookup(
&mut self,
channel: Channel
) -> Option<ChannelData<'_, dyn ProtocolObj>>

Look up what’s connected to channel (eg. the Protocol to which to forward).

fn att(&mut self) -> ChannelData<'_, AttributeServer<Self::AttributeProvider>>

Returns information about the Attribute Protocol on channel 0x0004.

Loading content...