Trait rubble::l2cap::ProtocolObj

pub trait ProtocolObj {
    // Required method
    fn process_message(
        &mut self,
        message: &[u8],
        responder: Sender<'_>
    ) -> Result<(), Error>;
}

Trait for protocols that sit on top of L2CAP (object-safe part).

A protocol can be connected to an L2CAP channel via a ChannelMapper.

Required Methods

fn process_message( &mut self, message: &[u8], responder: Sender<'_> ) -> Result<(), Error>

Process a message sent to the protocol.

The message is reassembled by L2CAP already, and the responder is guaranteed to fit a protocol payload of at least Protocol::RSP_PDU_SIZE Bytes, as defined by the protocol.

Errors

This method should only return an error when a critical problem occurs that can not be recovered from and that can not be reported back to the connected device using the protocol. This means that only things like unrecoverable protocol parsing errors should return an error here.