Enum rubble::link::ad_structure::AdStructure

#[non_exhaustive]
pub enum AdStructure<'a> {
    Flags(Flags),
    ServiceUuids16(ServiceUuids<'a, Uuid16>),
    ServiceUuids32(ServiceUuids<'a, Uuid32>),
    ServiceUuids128(ServiceUuids<'a, Uuid128>),
    ServiceData16 {
        uuid: Uuid16,
        data: &'a [u8],
    },
    CompleteLocalName(&'a str),
    ShortenedLocalName(&'a str),
    ManufacturerSpecificData {
        company_identifier: CompanyId,
        payload: &'a [u8],
    },
    Unknown {
        ty: u8,
        data: &'a [u8],
    },
}

A list of AD structures can be sent along with an advertising packet or scan response.

This mechanism allows a scanner to, for example, receive the device’s name without having to establish a connection.

Unless otherwise noted, each variant of this enum should only be included at most once per packet sent.

From a very unrepresentative scan, most devices seem to include Flags and Manufacturer Data, and optionally a device name, of course.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Flags(Flags)

Device flags and baseband capabilities.

This should be sent if any flags apply to the device. If not (ie. the value sent would be 0), this may be omitted.

Must not be used in scan response data.

ServiceUuids16(ServiceUuids<'a, Uuid16>)

ServiceUuids32(ServiceUuids<'a, Uuid32>)

ServiceUuids128(ServiceUuids<'a, Uuid128>)

ServiceData16

Fields

uuid: Uuid16

The 16-bit service UUID.

data: &'a [u8]

The associated service data. May be empty.

Service data with 16-bit service UUID.

CompleteLocalName(&'a str)

Sets the full (unabbreviated) device name.

This will be shown to the user when this device is found.

ShortenedLocalName(&'a str)

Sets the shortened device name.

ManufacturerSpecificData

Fields

company_identifier: CompanyId

payload: &'a [u8]

Set manufacturer specific data

Unknown

Fields

ty: u8

Type byte.

data: &'a [u8]

Raw data transmitted after the type.

An unknown or unimplemented AD structure stored as raw bytes.