Module rubble::link::advertising

Advertising channel operations.

This module defines PDUs, states and fields used by packets transmitted on the advertising channels. Generally, this includes everything needed to advertise as and scan for slave devices and to establish connections.

Note that while the types in here do not completely eliminate illegal values to be created, they do employ a range of sanity checks that prevent bogus packets from being sent by the stack.

Structs

ConnectRequestData

Connection parameters sent along with a ConnectRequest PDU (also known as LLData).

Header

16-bit Advertising Channel PDU header preceding the Payload.

PduBuf

Stores an advertising channel PDU.

Enums

Pdu

A parsed advertising channel PDU.

PduType

4-bit PDU type in Header.

SleepClockAccuracy

Indicates the master’s sleep clock accuracy (SCA) in ppm (parts per million).

Constants

ACCESS_ADDRESS

Access Address to use for all advertising channel packets.

CRC_PRESET

CRC initialization value for advertising channel packets.

MAX_PAYLOAD_SIZE

Max. advertising PDU payload size in Bytes.

RXADD_MASK

TXADD_MASK