Struct rubble::link::advertising::ConnectRequestData

pub struct ConnectRequestData {
    access_address: Hex<u32>,
    crc_init: Hex<u32>,
    win_size: Duration,
    win_offset: Duration,
    interval: Duration,
    latency: u16,
    timeout: Duration,
    chm: ChannelMap,
    hop: u8,
    sca: SleepClockAccuracy,
}

Connection parameters sent along with a ConnectRequest PDU (also known as LLData).

Fields
access_address: Hex<u32>crc_init: Hex<u32>win_size: Duration

Transmit window size in µs.

win_offset: Duration

Transmit window offset in µs.

interval: Duration

Connection interval in µs.

latency: u16

Slave latency (number of connection events).

timeout: Duration

Connection timeout.

chm: ChannelMaphop: u8sca: SleepClockAccuracy
Implementations

impl ConnectRequestData

pub fn access_address(&self) -> u32

Returns the Access Address to use for data channel communication.

The address is randomly generated by the initiator (the device sending the connection request) according to the requirements in the Bluetooth specification.

pub fn crc_init(&self) -> u32

Returns the initialization value for the CRC calculation.

The CRC polynomial is always the same.

pub fn channel_map(&self) -> &ChannelMap

Returns the channel map specified by the initiator.

pub fn hop(&self) -> u8

Returns the channel hop distance.

This must be in range 5..=16.

pub fn end_of_tx_window(&self) -> Duration

Returns the end of the transmit window from reception of the CONNECT_REQ containing self.

pub fn interval(&self) -> Duration

Returns the connection event interval in µs.

pub fn slave_latency(&self) -> u16

Returns the slave latency (as the number of connection events).

pub fn supervision_timeout(&self) -> Duration

Returns the connection supervision timeout (connSupervisionTimeout) to use for this connection.

If no data packet is received for this duration, the connection should be considered lost.

pub fn sca(&self) -> SleepClockAccuracy