Struct rubble::link::channel_map::ChannelMap

pub struct ChannelMap {
    raw: [u8; 5],
    num_used_channels: u8,
}

A map marking data channels as used or unused.

A channel map must mark at least 2 channels as used.

Fields
raw: [u8; 5]num_used_channels: u8
Implementations

impl ChannelMap

pub fn from_raw(raw: [u8; 5]) -> Self

Create a new channel map from the raw format used in Connection Request PDUs (ChM).

The first byte (LSB) contains flags for data channels 0 to 7, where the least significant bit is the flag for channel 0, and so on.

Since there are only 37 data channels, but 40 bits in the 5 Bytes, the 3 most significant bits in the last Byte of raw are considered reserved for future use (RFU) and are ignored by this function.

pub fn to_raw(&self) -> [u8; 5]

Returns the raw bytes encoding this channel map.

pub fn with_all_channels() -> Self

Creates a new channel map that marks all data channels as used.

pub fn num_used_channels(&self) -> u8

Returns the number of data channels marked as used by this map.

pub fn is_used(&self, channel: DataChannel) -> bool

Returns whether the given data channel is marked as used.

pub fn iter_used<'a>(&'a self) -> impl Iterator<Item = DataChannel> + 'a

Returns an iterator over all data channels marked as used in this map.

pub fn by_index(&self, n: u8) -> DataChannel

Returns the nth channel marked as used.

Panics

This will panic when n >= self.num_used_channels().