Module rubble::link::filter

Link-Layer Device Filtering.

Structs

AdvFilter
Advertising filter policy. Governs which devices may scan and connect to an advertising device.

AllowAll
An AddressFilter that allows all devices (ie. no whitelist in use).

ScanFilter
Scanner filter policy. Governs which devices will be scanned by this device.

WhitelistFilter
An AddressFilter that checks device addresses against a whitelist.

Traits

AddressFilter

Type Definitions

SingleIter

SliceIter