Struct rubble::link::filter::WhitelistFilter

pub struct WhitelistFilter<I: Iterator<Item = DeviceAddress> + Clone> { /* fields omitted */ }

An AddressFilter that checks device addresses against a whitelist.

This is a software filter, which allows checking against arbitrarily many device addresses, but might be less efficient than a hardware-based filter.

Implementations

impl<I: Iterator<Item = DeviceAddress> + Clone> WhitelistFilter<I>

pub fn new(allowed_addresses: I) -> Self

Creates a device whitelist from an iterator yielding the allowed device addresses.

The filter will clone and iterate over allowed_addresses for each incoming packet that needs to be checked against the filter.

impl<'a> WhitelistFilter<SliceIter<'a>>

pub fn from_slice(addresses: &'a [DeviceAddress]) -> Self

Creates a device whitelist from a slice of device addresses.

This is a convenience method provided to simplify using a slice as a device address whitelist.

impl WhitelistFilter<SingleIter>

pub fn from_address(address: DeviceAddress) -> Self

Creates a device whitelist that will allow a single device.