Module rubble::link::llcp

Defines packet structures used by the Link Layer Control Protocol.

Structs

ChannelMapReq

ConnectionParamRequest

A connection parameter update request or response (LL_CONNECTION_PARAM_REQ/ LL_CONNECTION_PARAM_RSP).

ConnectionUpdateData

Data transmitted with an LL_CONNECTION_UPDATE_REQ Control PDU, containing a new set of connection parameters.

Enums

ControlOpcode

Enumeration of all known LL Control PDU opcodes (not all of which might be supported).

ControlPdu

A structured representation of an LL Control PDU used by the Link Layer Control Protocol (LLCP).

VersionNumber

Enumeration of all possible VersNr for LL_VERSION_IND PDUs.