Struct rubble::link::llcp::ConnectionParamRequest

pub struct ConnectionParamRequest { /* fields omitted */ }

A connection parameter update request or response (LL_CONNECTION_PARAM_REQ/ LL_CONNECTION_PARAM_RSP).

Implementations

impl ConnectionParamRequest

pub fn new() -> Self

Creates a new connection update request structure filled with default values.

The returned structure will use conservative (maximally permissive) default values that will not usually result in a change in connection parameters, so users of this function likely want to call a setter afterwards.

pub fn set_conn_interval(&mut self, min: Duration, max: Duration)

Sets the minimum and maximum requested connection interval.

Parameters

min: Minimum connection interval to request.

max: Maximum connection interval to request.

Both min and max must be in range 7.5ms to 4s, or they will be constrained to lie in that range.

Both min and max will be rounded down to units of 1.25 ms.

Panics

This will panic if min > max.

pub fn min_conn_interval(&self) -> Duration

Returns the minimum requested connection interval.

pub fn max_conn_interval(&self) -> Duration

Returns the maximum requested connection interval.

pub fn slave_latency(&self) -> u16

Returns the slave latency in number of connection events.

pub fn supervision_timeout(&self) -> Duration

Returns the supervision timeout.