Module rubble::link::queue

An SPSC queue for data channel PDUs.

Data channel PDUs are received and transmitted in time-critical code, so they’re sent through a queue to be processed at a later time (eg. in the application’s idle loop).

This module defines the following important items:

The PacketQueue trait, which is implemented by all types providing a packet queue. A type implementing this trait is needed to use the Rubble stack.

The Producer and Consumer traits, which define the queue functionality used after splitting a PacketQueue.

The SimpleQueue, SimpleProducer and SimpleConsumer types, a minimal implementation of the queue interface defined by PacketQueue, Producer and Consumer.

Structs

Consume
Bundles a T along with information telling a queue whether to consume a packet.

SimpleConsumer
Consumer (reader) half returned by SimpleQueue::split.

SimpleProducer
Producer (writer) half returned by SimpleQueue::split.

SimpleQueue
A simple packet queue that can hold a single packet.

Traits

Consumer
The consuming (reading) half of a packet queue.

PacketQueue
A splittable SPSC queue for data channel PDUs.

Producer
The producing (writing) half of a packet queue.

Functions

run_tests
Runs Rubble’s packet queue testsuite against the given PacketQueue.