Struct rubble::link::Responder

pub struct Responder<M: ChannelMapper> { /* fields omitted */ }

Data channel packet processor.

This hooks up to the Real-Time part of the LE Link Layer via a packet queue. This part can run at a lower priority (eg. being driven in the apps idle loop) and receives and transmits packets using the packet queue.

Some LL Control PDUs sent as part of the Link Layer Control Protocol (LLCP) are answered by the responder directly, and all L2CAP data is forwarded to an L2CAPState<M>. Note that most LLCPDUs are handled directly by the real-time code.

Methods

impl<M: ChannelMapper> Responder<M>

pub fn new(tx: Producer, rx: Consumer, l2cap: L2CAPState<M>) -> Self

pub fn has_work(&mut self) -> bool

Returns true when this responder has work to do.

If this returns true, process may be called to process incoming packets and send outgoing ones.

pub fn process_one(&mut self) -> Result<(), Error>

Processes a single incoming packet in the packet queue.

Returns Error::Eof if there are no incoming packets in the RX queue.