Module rubble::phy

Physical layer interactions.

Don't expect to see much code here: Most of this layer is hardware.

BLE data is transmitted on 40 different RF channels numbered from 0 to 39 with ascending frequencies. Channels 0, 12 and 39 are reserved for advertising, all others are used for data transmission. BLE internally uses so-called "Channel Indices" which reorder the RF channels so that indices 0..=36 refer to data channels and 37..=39 refer to the advertising channels (presumably to simplify channel hopping). The Link-Layer is only interested in these channel indices, so only those are implemented here.

Structs

AdvertisingChannel
One of the three advertising channels (channel indices 37, 38 or 39).

DataChannel
One of 37 data channels on which data channel PDUs are sent between connected devices.

Traits

Radio
Trait for raw 2.4 GHz non-BLE-specific radios.