Struct rubble::time::Duration

pub struct Duration(_);

A duration with microsecond resolution.

This can represent a maximum duration of about 1 hour. Overflows will result in a panic, but shouldn't happen since the BLE stack doesn't deal with durations that large.

Implementations

impl Duration

pub const T_IFS: Self

The duration of the interframe spacing between BLE packets.

pub fn from_micros(micros: u32) -> Self

Creates a Duration from a number of microseconds.

pub fn from_millis(millis: u16) -> Self

Creates a Duration representing the given number of milliseconds.

pub fn from_secs(secs: u16) -> Self

Creates a Duration representing a number of seconds.

pub fn whole_secs(&self) -> u32

Returns the number of whole seconds that fit in self.

pub fn whole_millis(&self) -> u32

Returns the number of whole milliseconds that fit in self.

pub fn as_micros(&self) -> u32

Returns the number of microseconds represented by self.

pub fn subsec_micros(&self) -> u32

Returns the fractional part of microseconds in self.