Module rubble::uuid

BLE UUIDs (16, 32 or 128 bits).

Bluetooth assigns UUIDs to identify services and characteristics. In order to save space, many common UUIDs can be represented and transmitted as 16- or 32-bit aliases instead of the full 128 bits.

The shorter UUIDs can be converted to their full 128-bit counterparts by making use of the Bluetooth Base UUID, which is defined as 00000000-0000-1000-8000-00805F9B34FB.

A 16-bit UUID alias can be converted to its 32-bit equivalent by zero-extending it: 0xABCD becomes 0x0000ABCD.

A 32-bit UUID alias can then be converted to its full 128-bit equivalent by placing it in the first 4 Bytes of the Base UUID. Hence 0x1234ABCD would become 1234ABCD-0000-1000-8000-00805F9B34FB.

Structs

Uuid16

A 16-bit UUID alias.

Uuid32

A 32-bit UUID alias.

Uuid128

A full 128-bit UUID.

Enums

UuidKind

List of the supported UUID types.

Traits

IsUuid

Marker for UUID types.