Struct rubble::uuid::Uuid

pub struct Uuid(_);

A Universally Unique Identifier (UUID).

Methods

impl Uuid

pub fn to_hyphenated(self) -> Hyphenated

Creates a Hyphenated instance from a Uuid.

pub fn to_hyphenated_ref(&self) -> HyphenatedRef

Creates a HyphenatedRef instance from a Uuid reference.

pub fn to_simple(self) -> Simple

Creates a Simple instance from a Uuid.

pub fn to_simple_ref(&self) -> SimpleRef

Creates a SimpleRef instance from a Uuid reference.

pub fn to_urn(self) -> Urn

Creates a Urn instance from a Uuid.

pub fn to_urn_ref(&self) -> UrnRef

Creates a UrnRef instance from a Uuid reference.

impl Uuid

pub const NAMESPACE_DNS: Uuid

Uuid namespace for Domain Name System (DNS).

pub const NAMESPACE_OID: Uuid

Uuid namespace for ISO Object Identifiers (OIDs).

pub const NAMESPACE_URL: Uuid

Uuid namespace for Uniform Resource Locators (URLs).

pub const NAMESPACE_X500: Uuid

Uuid namespace for X.500 Distinguished Names (DNs).

pub fn nil() -> Uuid

The 'nil UUID'.

The nil UUID is special form of UUID that is specified to have all 128 bits set to zero, as defined in IETF RFC 4122 Section 4.1.7.

Examples

Basic usage:

use uuid::Uuid;

let uuid = Uuid::nil();

assert_eq!(
    uuid.to_hyphenated().to_string(),
    "00000000-0000-0000-0000-000000000000"
);

pub fn from_fields(
d1: u32,
d2: u16,
d3: u16,
d4: &[u8]
) -> Result<Uuid, BytesError>

Creates a Uuid from four field values in big-endian order.

Errors

This function will return an error if d4's length is not 8 bytes.

Examples

Basic usage:

use uuid::Uuid;

let d4 = [12, 3, 9, 56, 54, 43, 8, 9];

let uuid = Uuid::from_fields(42, 12, 5, &d4);
let uuid = uuid.map(|uuid| uuid.to_hyphenated().to_string());

let expected_uuid =
    Ok(String::from("0000002a-000c-0005-0c03-0938362b0809"));

assert_eq!(expected_uuid, uuid);

An invalid length:

use uuid::prelude::*;

let d4 = [12];

let uuid = uuid::Uuid::from_fields(42, 12, 5, &d4);

let expected_uuid = Err(uuid::BytesError::new(8, d4.len()));

assert_eq!(expected_uuid, uuid);

pub fn from_fields_le(
d1: u32,
d2: u16,
d3: u16,
d4: &[u8]
) -> Result<Uuid, BytesError>

Creates a Uuid from four field values in little-endian order.

The bytes in the d1, d2 and d3 fields will be converted into big-endian order.

Examples

use uuid::Uuid;

let d1 = 0xAB3F1097u32;
let d2 = 0x501Eu16;
let d3 = 0xB736u16;
let d4 = [12, 3, 9, 56, 54, 43, 8, 9];

let uuid = Uuid::from_fields_le(d1, d2, d3, &d4);
let uuid = uuid.map(|uuid| uuid.to_hyphenated().to_string());

let expected_uuid =
    Ok(String::from("97103fab-1e50-36b7-0c03-0938362b0809"));

assert_eq!(expected_uuid, uuid);

pub fn from_slice(b: &[u8]) -> Result<Uuid, BytesError>

Creates a Uuid using the supplied big-endian bytes.

Errors

This function will return an error if b has any length other than 16.

Examples

Basic usage:

use uuid::Uuid;

let bytes = [4, 54, 67, 12, 43, 2, 98, 76, 32, 50, 87, 5, 1, 33, 43, 87];

let uuid = Uuid::from_slice(&bytes);
let uuid = uuid.map(|uuid| uuid.to_hyphenated().to_string());

let expected_uuid =
    Ok(String::from("0436430c-2b02-624c-2032-570501212b57"));

assert_eq!(expected_uuid, uuid);

An incorrect number of bytes:

use uuid::prelude::*;

let bytes = [4, 54, 67, 12, 43, 2, 98, 76];

let uuid = Uuid::from_slice(&bytes);

let expected_uuid = Err(uuid::BytesError::new(16, 8));

assert_eq!(expected_uuid, uuid);

pub fn from_bytes(bytes: [u8; 16]) -> Uuid

Creates a Uuid using the supplied big-endian bytes.

pub fn from_random_bytes(bytes: [u8; 16]) -> Uuid

Deprecated since 0.7.2:
please use the uuid::Builder instead to set the variant and version

Creates a v4 Uuid from random bytes (e.g. bytes supplied from Rand crate)

Examples

Basic usage:

use uuid::Bytes;
use uuid::Uuid;

let bytes: Bytes = [
    70, 235, 208, 238, 14, 109, 67, 201, 185, 13, 204, 195, 90, 145, 63, 62,
];
let uuid = Uuid::from_random_bytes(bytes);
let uuid = uuid.to_hyphenated().to_string();

let expected_uuid = String::from("46ebd0ee-0e6d-43c9-b90d-ccc35a913f3e");

assert_eq!(expected_uuid, uuid);

pub fn get_variant(&self) -> Option<Variant>

Returns the variant of the Uuid structure.

This determines the interpretation of the structure of the UUID. Currently only the RFC4122 variant is generated by this module.

Variant Reference

pub fn get_version_num(&self) -> usize

Returns the version number of the Uuid.

This represents the algorithm used to generate the contents.

Currently only the Random (V4) algorithm is supported by this module. There are security and privacy implications for using older versions - see Wikipedia: Universally Unique Identifier for details.

Version Reference

pub fn get_version(&self) -> Option<Version>

Returns the version of the Uuid.

This represents the algorithm used to generate the contents

pub fn as_fields(&self) -> (u32, u16, u16, &[u8; 8])

Returns the four field values of the UUID in big-endian order.

These values can be passed to the from_fields() method to get the original Uuid back.

The first field value represents the first group of (eight) hex digits, taken as a big-endian u32 value. For V1 UUIDs, this field represents the low 32 bits of the timestamp.

The second field value represents the second group of (four) hex digits, taken as a big-endian u16 value. For V1 UUIDs, this field represents the middle 16 bits of the timestamp.

The third field value represents the third group of (four) hex digits, taken as a big-endian u16 value. The 4 most significant bits give the UUID version, and for V1 UUIDs, the last 12 bits represent the high 12 bits of the timestamp.

The last field value represents the last two groups of four and twelve hex digits, taken in order. The first 1-3 bits of this indicate the UUID variant, and for V1 UUIDs, the next 13-15 bits indicate the clock sequence and the last 48 bits indicate the node ID.

Examples

use uuid::Uuid;

let uuid = Uuid::nil();
assert_eq!(uuid.as_fields(), (0, 0, 0, &[0u8; 8]));

let uuid = Uuid::parse_str("936DA01F-9ABD-4D9D-80C7-02AF85C822A8").unwrap();
assert_eq!(
    uuid.as_fields(),
    (
        0x936DA01F,
        0x9ABD,
        0x4D9D,
        b"\x80\xC7\x02\xAF\x85\xC8\x22\xA8"
    )
);

pub fn to_fields_le(&self) -> (u32, u16, u16, &[u8; 8])

Returns the four field values of the UUID in little-endian order.

The bytes in the returned integer fields will be converted from big-endian order.

Examples

use uuid::Uuid;

let uuid = Uuid::parse_str("936DA01F-9ABD-4D9D-80C7-02AF85C822A8").unwrap();
assert_eq!(
    uuid.to_fields_le(),
    (
        0x1FA06D93,
        0xBD9A,
        0x9D4D,
        b"\x80\xC7\x02\xAF\x85\xC8\x22\xA8"
    )
);

pub fn as_bytes(&self) -> &[u8; 16]

Returns an array of 16 octets containing the UUID data. This method wraps [Uuid::as_bytes]

pub fn to_timestamp(&self) -> Option<(u64, u16)>

Returns an Optional Tuple of (u64, u16) representing the timestamp and counter portion of a V1 UUID. If the supplied UUID is not V1, this will return None

pub fn parse_str(input: &str) -> Result<Uuid, ParseError>

Parses a Uuid from a string of hexadecimal digits with optional hyphens.

Any of the formats generated by this module (simple, hyphenated, urn) are supported by this parsing function.

pub fn is_nil(&self) -> bool

Tests if the UUID is nil

pub fn encode_buffer() -> [u8; 45]

A buffer that can be used for encode_... calls, that is guaranteed to be long enough for any of the adapters.

Examples

use uuid::Uuid;

let uuid = Uuid::nil();

assert_eq!(
    uuid.to_simple().encode_lower(&mut Uuid::encode_buffer()),
    "00000000000000000000000000000000"
);

assert_eq!(
    uuid.to_hyphenated()
        .encode_lower(&mut Uuid::encode_buffer()),
    "00000000-0000-0000-0000-000000000000"
);

assert_eq!(
    uuid.to_urn().encode_lower(&mut Uuid::encode_buffer()),
    "urn:uuid:00000000-0000-0000-0000-000000000000"
);