Trait rubble::uuid::IsUuid

pub trait IsUuid: for<'a> FromBytes<'a> + ToBytes + Copy {
    const KIND: UuidKind;
}

Marker for UUID types.

This is useful when being generic over the specific type of UUID used. It also brings in the ToBytes and FromBytes trait bounds that are likely needed as well.

Required Associated Constants

const KIND: UuidKind