Trait rubble::att::AttributeProvider

pub trait AttributeProvider {
    fn for_attrs_in_range(
        &mut self, 
        range: HandleRange, 
        f: impl FnMut(&Self, Attribute) -> Result<(), Error>
    ) -> Result<(), Error>;

    fn is_grouping_attr(&self, uuid: AttUuid) -> bool;

    fn group_end(&self, handle: Handle) -> Option<&Attribute>;
}

Trait for attribute sets that can be hosted by an AttributeServer.

Required methods

fn for_attrs_in_range(
&mut self,
range: HandleRange,
f: impl FnMut(&Self, Attribute) -> Result<(), Error>
) -> Result<(), Error>

Calls a closure f with every attribute whose handle is inside range, ascending.

If f returns an error, this function will stop calling f and propagate the error upwards. If f returns Ok(()), iteration will continue.

This function would ideally return an iterator instead of invoking a callback, but it's not currently possible to express the iterator type generically (it would need lifetime-generic associated types), and all workarounds seem to be severely limiting.

fn is_grouping_attr(&self, uuid: AttUuid) -> bool

Returns whether uuid is a valid grouping attribute type that can be used in Read By Group Type requests.

fn group_end(&self, handle: Handle) -> Option<&Attribute>

Queries the last attribute that is part of the attribute group denoted by the grouping attribute at handle.

If handle does not refer to a grouping attribute, returns None.

Groups themselves are collections of attributes. An attribute is in exactly zero or one groups.

For primary services and secondary services, the BLE specification gives exact definitions of what's in the group. The group begins with the "primary service" or "secondary service" attribute, is followed by the various attributes contained within that service, and ends with the last attribute contained within that service.

TODO: document what the BLE spec has to say about grouping for characteristics.

Loading content...