Trait rubble::att::AttributeProvider

pub trait AttributeProvider {
    fn for_each_attr(
        &mut self, 
        f: &mut dyn FnMut(&mut Attribute) -> Result<(), Error>
    ) -> Result<(), Error>;

    fn is_grouping_attr(&self, uuid: AttUuid) -> bool;

    fn group_end(&self, handle: AttHandle) -> Option<&Attribute>;

    fn any(&mut self, filter: &mut dyn FnMut(&mut Attribute) -> bool) -> bool { ... }
}

Trait for attribute sets that can be hosted by an AttributeServer.

Required methods

fn for_each_attr(
&mut self,
f: &mut dyn FnMut(&mut Attribute) -> Result<(), Error>
) -> Result<(), Error>

Calls a closure f with every attribute stored in self.

All attributes will have ascending, consecutive handle values starting at 0x0001.

fn is_grouping_attr(&self, uuid: AttUuid) -> bool

Returns whether uuid is a valid grouping attribute that can be used in Read By Group Type requests.

fn group_end(&self, handle: AttHandle) -> Option<&Attribute>

Queries the last attribute that is part of the attribute group denoted by the grouping attribute at handle.

If handle does not refer to a grouping attribute, returns None.

TODO: Human-readable docs that explain what grouping is

Loading content...
Provided methods

fn any(&mut self, filter: &mut dyn FnMut(&mut Attribute) -> bool) -> bool

Returns whether the filter closure matches any attribute in self.

Loading content...