Struct rubble::beacon::BeaconScanner

pub struct BeaconScanner<C: ScanCallback, F: AddressFilter> {
    cb: C,
    filter: ScanFilter<F>,
    interval: Duration,
    channel: AdvertisingChannel,
}

A passive scanner for non-connectable beacon advertisements.

Fields
cb: Cfilter: ScanFilter<F>interval: Durationchannel: AdvertisingChannel
Implementations

impl<C: ScanCallback> BeaconScanner<C, AllowAll>

pub fn new(callback: C) -> Self

Creates a BeaconScanner that will report beacons from any device.

impl<C: ScanCallback, F: AddressFilter> BeaconScanner<C, F>

pub fn with_filter(callback: C, scan_filter: F) -> Self

Creates a BeaconScanner with a custom device filter.

pub fn configure(&mut self, now: Instant, interval: Duration) -> Cmd

Configures the BeaconScanner and returns a Cmd to apply to the radio.

The next_update field of the returned Cmd specifies when to call timer_update the next time. The timer used for this does not have to be very accurate, it is only used to switch to the next advertising channel after interval elapses.

pub fn timer_update(&mut self, now: Instant) -> Cmd

Updates the BeaconScanner after the configured timer has fired.

This switches to the next advertising channel and will listen there.

pub fn process_adv_packet(
&mut self,
header: Header,
payload: &[u8],
crc_ok: bool
) -> Cmd

Processes a received advertising channel packet.

This should be called whenever the radio receives a packet on the configured advertising channel.