Struct rubble::l2cap::BleChannelMap

pub struct BleChannelMap<A: AttributeProvider, S: SecurityLevel> { /* fields omitted */ }

A fixed BLE channel map that provides only the required channel endpoints and does not allow dynamic channels.

The channels are mapped as follows (no other channels are supported):

0x0004: Attribute protocol (ATT).

0x0005: LE L2CAP signaling channel.

0x0006: LE Security Manager protocol.

Implementations

impl BleChannelMap<NoAttributes, NoSecurity>

pub fn empty() -> Self

Creates a new channel map with no backing data for the connected protocols.

This means:

The attribute server on channel 0x0004 will host an empty attribute set.

The security manager on channel 0x0006 will not support pairing or any security.

impl<A: AttributeProvider> BleChannelMap<A, NoSecurity>

pub fn with_attributes(att: A) -> Self

pub fn attribute_provider(&mut self) -> &mut A

Provides mutable access to the underlying AttributeProvider.