Trait rubble::l2cap::Protocol

pub trait Protocol: ProtocolObj {
    const RSP_PDU_SIZE: u8;
}

Trait for protocols that sit on top of L2CAP (non-object-safe part).

This extends the ProtocolObj trait with other protocol properties.

Associated Constants

const RSP_PDU_SIZE: u8

Minimum size needed by PDUs sent by this protocol.

Incoming PDUs will only be forwarded to the protocol if there is at least this much space in the TX buffer.