Struct rubble::link::advertising::Header

pub struct Header(_);

16-bit Advertising Channel PDU header preceding the Payload.

The header looks like this:

LSB                                                                     MSB
+------------+------------+---------+---------+--------------+------------+
|  PDU Type  |     -      |  TxAdd  |  RxAdd  |    Length    |     -      |
|  (4 bits)  |  (2 bits)  | (1 bit) | (1 bit) |   (6 bits)   |  (2 bits)  |
+------------+------------+---------+---------+--------------+------------+

The TxAdd and RxAdd field are only used for some payloads, for all others, they should be set to 0.

Length may be in range 6 to 37 (inclusive). With the 2-Byte header this is exactly the max. on-air packet size.

Implementations

impl Header

pub fn new(ty: PduType) -> Self

Creates a new Advertising Channel PDU header specifying the Payload type ty.

pub fn parse(raw: &[u8]) -> Self

pub fn to_u16(&self) -> u16

Returns the raw representation of the header.

The returned u16 must be transmitted LSb first as the first 2 octets of the PDU.

pub fn type_(&self) -> PduType

Returns the PDU type specified in the header.

pub fn tx_add(&self) -> bool

Returns the state of the TxAdd field.

pub fn set_tx_add(&mut self, value: bool)

Sets the TxAdd field's value.

pub fn rx_add(&self) -> bool

Returns the state of the RxAdd field.

pub fn set_rx_add(&mut self, value: bool)

Sets the RxAdd field's value.

pub fn payload_length(&self) -> u8

Returns the length of the payload in octets as specified in the Length field.

According to the spec, the length must be in range 6...37, but this isn't checked by this function.

pub fn set_payload_length(&mut self, length: u8)

Sets the payload length of this PDU.

The length must be in range 6...37, otherwise this function panics.