Struct rubble::link::LinkLayer

pub struct LinkLayer<C: Config> {
    dev_addr: DeviceAddress,
    state: State<C>,
    timer: C::Timer,
}

Implementation of the real-time BLE Link-Layer logic.

Users of this struct must provide an interface to the platform’s hardware by implementing Config.

Fields
dev_addr: DeviceAddressstate: State<C>timer: C::Timer
Implementations

impl<C: Config> LinkLayer<C>

pub fn new(dev_addr: DeviceAddress, timer: C::Timer) -> Self

Creates a new Link-Layer.

Parameters

dev_addr: The device address to broadcast as.

timer: A Timer implementation.

tx: Input queue of packets to transmit when connected.

rx: Output queue of received packets when connected.

pub fn timer(&mut self) -> &mut C::Timer

Returns a reference to the timer instance used by the Link-Layer.

pub fn start_advertise(
&mut self,
interval: Duration,
data: &[AdStructure<'_>],
transmitter: &mut C::Transmitter,
tx: <<C as Config>::PacketQueue as PacketQueue>::Consumer,
rx: <<C as Config>::PacketQueue as PacketQueue>::Producer
) -> Result<NextUpdate, Error>

Starts advertising this device, optionally sending data along with the advertising PDU.

pub fn process_adv_packet(
&mut self,
rx_end: Instant,
tx: &mut C::Transmitter,
header: Header,
payload: &[u8],
crc_ok: bool
) -> Cmd

Process an incoming packet from an advertising channel.

The access address of the packet must be ADVERTISING_ADDRESS.

Parameters

rx_end: A timestamp indicating when the packet was fully received.

tx: A packet transmitter.

header: The header of the received packet.

payload: The packet payload following the header.

crc_ok: Whether the packet’s CRC is correct.

pub fn process_data_packet(
&mut self,
rx_end: Instant,
tx: &mut C::Transmitter,
header: Header,
payload: &[u8],
crc_ok: bool
) -> Cmd

Process an incoming data channel packet.

pub fn update_timer(&mut self, tx: &mut C::Transmitter) -> Cmd

Update the Link-Layer state after the timer expires.

This should be called whenever the timer set by the last returned Cmd has expired.

Parameters

tx: A Transmitter for sending packets.

pub fn connection(&self) -> Option<&Connection<C>>

Returns a reference to the connection state.

If the Link Layer is not currently in a connection, returns None.

pub fn is_advertising(&self) -> bool

Returns whether the Link-Layer is currently broadcasting advertisement packets.

pub fn is_connected(&self) -> bool

Returns whether the Link-Layer is currently connected.